+++ packages/matter.js/src/CommissioningServer.ts
@@ -729,20 +729,9 @@ export class CommissioningServer extends MatterNode {
         return this.deviceInstance?.getFabrics().map(fabric => fabric.getExternalInformation()) ?? [];
     }
 
-    /** get some basic details of all Fabrics where an active subscription exists. */
-    getSubscribedFabricInformation() {
+    /** Get some basic details of all currently active sessions. */
+    getActiveSessionInformation() {
         if (!this.isCommissioned()) return [];
-        return this.interactionServer?.getSubscribedFabricInformation() ?? [];
-    }
-
-    /** Get the number of active subscriptions. */
-    getNumberOfActiveSubscriptions() {
-        if (!this.isCommissioned()) return 0;
-        return this.interactionServer?.getNumberOfActiveSubscriptions() ?? 0;
-    }
-
-    /** Returns true if at least one active subscription exists. */
-    isSubscribed() {
-        return this.getNumberOfActiveSubscriptions() > 0;
+        return this.deviceInstance?.getActiveSessionInformation() ?? [];
     }
 }