+++ packages/matter.js/src/protocol/interaction/InteractionServer.ts
@@ -628,14 +628,4 @@ export class InteractionServer implements ProtocolHandler<MatterDevice> {
             subscription.cancel();
         }
     }
-
-    getSubscribedFabricInformation() {
-        return Array.from(this.subscriptionMap.values()).map(subscription =>
-            subscription.getFabric().getExternalInformation(),
-        );
-    }
-
-    getNumberOfActiveSubscriptions() {
-        return this.subscriptionMap.size;
-    }
 }